% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_workshops.R
\name{build_workshops}
\alias{build_workshops}
\title{Render workshops}
\usage{
build_workshops(path = ".", download = FALSE, id = NULL,
  Rmdfiles = NULL, verbose = TRUE)
}
\arguments{
\item{path}{path where the workshops files are or a path to a folder where the files will be extracted if download is set to \code{TRUE}.}

\item{download}{a logical. Should the workshop source files be downloaded? Default set to \code{FALSE}, it \code{TRUE} then \code{id} must be specified.}

\item{id}{workshops identifier.}

\item{Rmdfiles}{R Markdown source files, if \code{NULL} (default) then}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Render the workshops and download the source files if requested.
}
