% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workshops.R
\name{clean_workshops}
\alias{clean_workshops}
\title{Clean up workshop folder}
\usage{
clean_workshops(path = ".", files = TRUE, cache = TRUE,
  verbose = TRUE)
}
\arguments{
\item{path}{path where the workshops files are to be cleaned up.}

\item{files}{a logical. Should workshop files be removed?}

\item{cache}{a logical. Should workshop cache be cleared?}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Remove workshop files and clear cache.
}
