% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_workshops.R
\name{download_workshops}
\alias{download_workshops}
\title{Download workshops}
\usage{
download_workshops(ids, path = ".", verbose = TRUE)
}
\arguments{
\item{ids}{workshop identifiers.}

\item{path}{path where the workshops files are or a path to a folder where the files will be extracted if download is set to \code{TRUE}.}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Download a selecttion of workshops.
}
\examples{
download_workshops(1)
}
