% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_Rcode_workshops.R
\name{extract_Rcode_workshops}
\alias{extract_Rcode_workshops}
\title{Extract R code}
\usage{
extract_Rcode_workshops(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{path to a folder containing one or more workshops.}

\item{verbose}{a logical. Should extra information be reported on
progress?}
}
\description{
Extract R code from workshops R Markdown files.
}
