% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_slides.R
\name{first_slides}
\alias{first_slides}
\title{Returns the two first slides of a workshops}
\usage{
first_slides(workshop_id, pkgs, lang = c("en", "fr"), clip = FALSE)
}
\arguments{
\item{workshop_id}{workshop identifier, see \code{\link[=workshops_df]{workshops_df()}}.}

\item{pkgs}{vector of packages required for the workshops, see \code{\link[=list_packages]{list_packages()}}.}

\item{lang}{the language for the badges (either "en" or "fr").}

\item{clip}{a logical. Should the badges be copied to the clipboard?}
}
\description{
Returns the two first slides of a workshop (title slide not included).
}
\examples{
first_slides(2, c("dplyr", "tidyr"))
}
