% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_badges.R
\name{get_badges}
\alias{get_badges}
\title{Return bagdes for a given workshop.}
\usage{
get_badges(workshop_id, lang = c("en", "fr"))
}
\arguments{
\item{workshop_id}{workshop identifier, see \code{\link[=workshops_df]{workshops_df()}}.}

\item{lang}{the language for the badges.}
}
\description{
Return bagdes for a given workshop.
}
\examples{
get_badges(1)
get_badges(4, lang = "fr")
}
