% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_workshops_pkgs.R
\name{install_workshops_pkgs}
\alias{install_workshops_pkgs}
\title{Read list of dependencies and install packages}
\usage{
install_workshops_pkgs(file = "pkgs.yaml", verbose = TRUE)
}
\arguments{
\item{file}{path to the list of dependencies}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Read list of dependencies and install packages
}
