% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{Create a list of packages}
\usage{
list_packages(pkgs, add_cls = TRUE, clip = TRUE, show = FALSE)
}
\arguments{
\item{pkgs}{vector of packages.}

\item{add_cls}{a logical. Should the command lines to install packages be added.}

\item{clip}{a logical. Should the package list be copied to the clipboard.}

\item{show}{a logical. Should the badges be displayed?}
}
\description{
In the first slides of a workshop the set of packages should be specified.
This function create a bullet points list with CRAN URL for packages as
well as a code chunk to install all packages.
}
\examples{
list_packages(c("cli", "glue"), clip = FALSE)
}
