% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_workshop.R
\name{mail_workshop}
\alias{mail_workshop}
\title{Generate presenter email}
\usage{
mail_workshop(
  workshop_id,
  pres_name,
  lang = c("both", "en", "fr"),
  details_en = "",
  details_fr = "",
  markdownify = TRUE
)
}
\arguments{
\item{workshop_id}{workshop identifier, see \code{\link[=workshops_df]{workshops_df()}}.}

\item{pres_name}{a character string with the presenter name.}

\item{lang}{the langage of the mail, by default mail will contain
a French and an English version.}

\item{details_en}{a character string indicating where and when the workshop will be held in English.}

\item{details_fr}{same as \code{details_en} but in French.}

\item{markdownify}{a logical. Should \url{balstula::md()} be used. Default
is set to \code{TRUE}.}
}
\description{
Generate presenter email to be sent two weeks beforehand.
}
\details{
Currently, the email is generated with \code{\link[blastula:compose_email]{blastula::compose_email()}} and
you need to copy paste it.
}
\examples{
mail_en <- mail_workshop(7, "Kevin Cazelles", lang = "en",
  details_en = "\u00e0 l'UdeM, salle A-2553,  campus MIL, le vendredi 28 F\u00e9vrier
  2010 de 13h \u00e0 17h")
}
