% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_workshops.R
\name{render_workshops}
\alias{render_workshops}
\title{Render workshops}
\usage{
render_workshops(files, verbose = TRUE)
}
\arguments{
\item{files}{path to the workshops' Rmd file(s) to be rendered.}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Convert R Markdown files to HTML presentations.
}
