% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{update_template}
\alias{update_template}
\title{Update template}
\usage{
update_template(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{path to the workshops where to find the template to be updated.}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Download the template, check whether the files must be updated and do the
update of required.
}
