% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_workshops.R
\name{build_workshops}
\alias{build_workshops}
\title{Render workshops}
\usage{
build_workshops(
  path = ".",
  download = FALSE,
  id = NULL,
  Rmdfiles = NULL,
  update_template = FALSE,
  pdf = FALSE,
  script = FALSE,
  upgrade = "never",
  verbose = TRUE
)
}
\arguments{
\item{path}{path where the workshops files are or a path to a folder where the files will be extracted if download is set to \code{TRUE}.}

\item{download}{a logical. Should the workshop source files be downloaded? Default set to \code{FALSE}, it \code{TRUE} then \code{id} must be specified.}

\item{id}{workshops identifier.}

\item{Rmdfiles}{R Markdown source files, if \code{NULL} (default) then, files will be searched for in the current directory.}

\item{update_template}{a logical. Should the template files be updated. Note that if these files are missing then they will be downloaded.}

\item{pdf}{a logical. Should a pdf version of the template be produced?}

\item{script}{a logical. Should the R script be extracted?}

\item{upgrade}{Should how of date packages be updated? One of "default",
"ask", "always", or "never". "default", see \code{\link[remotes:install_deps]{remotes::install_deps()}} for
further details.}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\description{
Render the workshops and download the source files if requested.
}
