% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_badges.R
\name{get_badges}
\alias{get_badges}
\title{Return badges for a given workshop.}
\usage{
get_badges(
  workshop_id,
  lang = c("en", "fr"),
  style = c("flat-square", "plastic", "flat", "for-the-badge", "social"),
  clip = TRUE,
  show = TRUE
)
}
\arguments{
\item{workshop_id}{workshop identifier, see \code{\link[=workshops_df]{workshops_df()}}.}

\item{lang}{the language for the badges (either "en" or "fr").}

\item{style}{badges style.}

\item{clip}{a logical. Should the badges be copied to the clipboard?}

\item{show}{a logical. Should the badges be displayed?}
}
\description{
Return badges for a given workshop.
}
\examples{
get_badges(1, clip = FALSE)
get_badges(4, lang = "fr", clip = FALSE)
}
