% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_workshops_pkgs.R
\name{install_workshops_pkgs}
\alias{install_workshops_pkgs}
\alias{install_workshop_pkgs_indiv}
\alias{install_cran_pkgs}
\alias{install_cran_archived_pkgs}
\alias{install_repos_pkgs}
\title{Read list of dependencies and install packages}
\usage{
install_workshops_pkgs(file = "pkgs.yaml", verbose = TRUE)

install_workshop_pkgs_indiv(file = "pkgs.yaml", verbose = TRUE)

install_cran_pkgs(x)

install_cran_archived_pkgs(x)

install_repos_pkgs(x)
}
\arguments{
\item{file}{path to the list of dependencies}

\item{verbose}{a logical. Should extra information be reported on progress?}

\item{x}{a character string of the name of a package.}
}
\description{
Read list of dependencies and install packages
}
