% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{qcbsRworkshops}
\alias{qcbsRworkshops}
\alias{qcbsRworkshops-package}
\title{qcbsRworkshops}
\description{
Set of helper functions to ease contribution to QCBS R workshops.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/QCBSRworkshops/qcbsRworkshops}
  \item \url{https://github.com/QCBSRworkshops/QCBSRworkshops.github.io}
  \item Report bugs at \url{https://github.com/QCBSRworkshops/qcbsRworkshops/issues}
}

}
\author{
\strong{Maintainer}: Kevin Cazelles \email{kcazelle@uoguelph.ca} (\href{https://orcid.org/0000-0001-6619-9874}{ORCID})

Authors:
\itemize{
  \item Marie-Hélène Brice (\href{https://orcid.org/0000-0003-0390-4833}{ORCID})
}

}
\keyword{internal}
